Hash Identifier Column

Client-side function that retrieves hashed identifiers from a DataSHIELD server. Used as the first step in record matching for vertically partitioned data.

Usage

ds.hashId(data_name, id_col, algo = "sha256", datasource = NULL)

Arguments

data_name

Character string. Name of the data frame on the server.

id_col

Character string. Name of the identifier column to hash.

algo

Character string. Hash algorithm. Default is "sha256".

datasource

A DataSHIELD connection object (DSConnection). If NULL, uses the first available connection.

Value

A list containing:

hashes: Character vector of hashed identifiers

n: Number of observations

Details

This function is typically called on the "reference" server in a vertical partitioning scenario. The returned hashes are then passed to ds.alignRecords on other servers to align the data.

See also

Examples

if (FALSE) { # \dontrun{
# Get hashes from reference server
ref_hashes <- ds.hashId("D", "patient_id", datasource = conns$server1)

# Use hashes to align other servers
ds.alignRecords("D", "patient_id", ref_hashes$hashes,
                datasources = conns[c("server2", "server3")])
} # }